image-verify/requirements.txt
flask>=3.0
pillow>=10.0
numpy>=1.26
opencv-python-headless>=4.9
python-magic>=0.4.27
requests>=2.31
